Change to int

+++ src/runtime/constants.h
@@ -8,15 +8,13 @@
  * public API of the runtime.
  */
 
-#include "HalideRuntime.h"
-
 namespace Halide {
 namespace Runtime {
 namespace Internal {
 namespace Constants {
 
 /** The threshold at which "stack" allocations should actually be backed by the heap. */
-static constexpr size_t maximum_stack_allocation_bytes = 16384;
+static constexpr int maximum_stack_allocation_bytes = 16384;
 
 }  // namespace Constants
 }  // namespace Internal
